+++ R/censobr_cache.R
@@ -51,6 +51,15 @@ censobr_cache <- function(list_files = TRUE,
     if (delete_file=='all') {
       unlink(files, recursive = TRUE)
       message(paste0("All files have been removed."))
+
+      ## also delete any files from old data releases
+        # determine old cache
+        dir_above <- dirname(censobr_env$cache_dir)
+        all_cache <- list.files(dir_above, pattern = 'censobr',full.names = TRUE)
+        old_cache <- all_cache[!grepl(censobr_env$pkgv, all_cache)]
+        # delete
+        unlink(old_cache, recursive = TRUE)
+
     }
   }