+++ R/onLoad.R
@@ -8,5 +8,6 @@ censobr_env <- new.env(parent = emptyenv()) # nocov start
 
   # local cache dir
   pkgv <- paste0('censobr_', censobr_env$data_release)
+  censobr_env$pkgv <- pkgv
   censobr_env$cache_dir <- tools::R_user_dir(pkgv, which = 'cache')
-  } # nocov end
+} # nocov end